Introduce content module
